moved get course/history from viviswap to dlt, updated tests
