+++ api_types/src/api/viviswap/mod.rs
@@ -1,6 +1,5 @@
 // since viviswap is quite big, we separate the objects into buckets
 pub mod contract;
-pub mod course;
 pub mod currency;
 pub mod detail;
 pub mod kyc;
