+++ sdk/src/backend/dlt.rs
@@ -3,11 +3,15 @@
 use super::error::{ApiError, Result};
 use crate::{core::config::Config, types::newtypes::AccessToken};
 use api_types::api::{
-    dlt::{AddressQueryParameters, ApiGetNetworksResponse, SetUserAddressRequest},
+    dlt::{
+        AddressQueryParameters, ApiGetNetworksResponse, GetCourseRequestQueries, GetCourseResponse,
+        SetUserAddressRequest,
+    },
     networks::ApiNetwork,
 };
 use log::{debug, error, info};
 use reqwest::StatusCode;
+use rust_decimal::Decimal;
 
 /// Puts the user crypto currency address in the backend
 ///
@@ -124,12 +128,66 @@ pub async fn get_networks(config: &Config, access_token: &AccessToken) -> Result
     }
 }
 
+/// Get exchange rate.
+///
+/// # Arguments
+///
+/// * `config` - The configuration object.
+/// * `access_token` - The access token for authentication.
+/// * `username` - The username of the user.
+///
+/// # Returns
+///
+/// Returns a `Result` containing the exchange rate as `f32` if successful.
+///
+/// # Errors
+///
+/// This function can return an `Error` if the request fails or if the response status is unauthorized.
+pub async fn get_exchange_rate(config: &Config, access_token: &AccessToken, network_key: String) -> Result<Decimal> {
+    info!("get_exchange_rate for network_key = {:?}", network_key);
+
+    let base_url = &config.backend_url;
+    let url = format!("{base_url}/courses");
+
+    let query = GetCourseRequestQueries { network_key };
+
+    let client = reqwest::Client::new();
+    let response = client
+        .get(&url)
+        .bearer_auth(access_token.as_str())
+        .header("X-APP-NAME", &config.auth_provider)
+        .query(&query)
+        .send()
+        .await?;
+    debug!("Response: {response:#?}");
+
+    match response.status() {
+        StatusCode::OK => {
+            let course = response.json::<GetCourseResponse>().await?.course;
+            Ok(course.course.0)
+        }
+        StatusCode::UNAUTHORIZED => Err(ApiError::MissingAccessToken),
+        _ => {
+            let status = response.status();
+            let text = response.text().await?;
+            error!(
+                "Failed to get node urls from backend: Response status: {}, Response text: {:?}",
+                status, text
+            );
+            Err(ApiError::UnexpectedResponse {
+                code: status,
+                body: text,
+            })
+        }
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
     use crate::testing_utils::{
-        example_api_network, set_config, ADDRESS, AUTH_PROVIDER, ETH_NETWORK_KEY, HEADER_X_APP_NAME, IOTA_NETWORK_KEY,
-        TOKEN,
+        example_api_network, example_exchange_rate_response, set_config, ADDRESS, AUTH_PROVIDER, ETH_NETWORK_KEY,
+        HEADER_X_APP_NAME, IOTA_NETWORK_KEY, TOKEN,
     };
     use mockito::Matcher;
 
@@ -222,6 +280,51 @@ mod tests {
         match expected {
             Ok(resp) => {
                 assert_eq!(response.unwrap(), resp.networks);
+            }
+            Err(ref expected_err) => {
+                assert_eq!(response.err().unwrap().to_string(), expected_err.to_string());
+            }
+        }
+        mock_server.assert();
+    }
+
+    #[rstest::rstest]
+    #[case(200, Ok(example_exchange_rate_response()))]
+    #[case(401, Err(ApiError::MissingAccessToken))]
+    #[case(500, Err(ApiError::UnexpectedResponse {
+        code: StatusCode::INTERNAL_SERVER_ERROR,
+        body: "".to_string() 
+    }))]
+    #[case(501, Err(ApiError::UnexpectedResponse {
+        code: StatusCode::NOT_IMPLEMENTED,
+        body: "".to_string() 
+    }))]
+    #[tokio::test]
+    async fn test_get_exchange_rate(#[case] status_code: usize, #[case] expected: Result<GetCourseResponse>) {
+        // Arrange
+        let (mut srv, config, _cleanup) = set_config().await;
+
+        let body = serde_json::to_string(&example_exchange_rate_response()).unwrap();
+
+        let mut mock_server = srv
+            .mock("GET", "/api/courses")
+            .match_header(HEADER_X_APP_NAME, AUTH_PROVIDER)
+            .match_header("authorization", format!("Bearer {}", TOKEN.as_str()).as_str())
+            .match_query(Matcher::Exact("network_key=IOTA".to_string()))
+            .with_status(status_code);
+        // Conditionally add the response body only for the 200 status code
+        if status_code == 200 {
+            mock_server = mock_server.with_body(&body);
+        }
+        let mock_server = mock_server.expect(1).create();
+
+        // Act
+        let response = get_exchange_rate(&config, &TOKEN, "IOTA".to_string()).await;
+
+        // Assert
+        match expected {
+            Ok(resp) => {
+                assert_eq!(response.unwrap(), resp.course.course.0);
             }
             Err(ref expected_err) => {
                 assert_eq!(response.err().unwrap().to_string(), expected_err.to_string());
